pyproject.toml
[project]
name = "currency-agent"
version = "0.1.0"
description = "Currency conversion agent using ADK, MCP and A2A"
readme = "README.md"
requires-python = ">=3.13, <3.14"
dependencies = [
    "a2a-sdk>=0.3.3",
    "fastmcp>=2.11.3",
    "google-adk>=1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
    "protobuf<6.0.0",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "pre-commit>=4.1.0",
    "yamllint>=1.35.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]
